Add unauthorized guard for the admin panel
