+++ front/packages/models/src/query.tsx
@@ -183,6 +183,7 @@ export type QueryPage<Props = {}, Items = unknown> = ComponentType<
 	getLayout?:
 		| QueryPage<{ page: ReactElement }>
 		| { Layout: QueryPage<{ page: ReactElement }>; props: object };
+	requiredPermissions?: string[]
 	randomItems?: Items[];
 };
