+++ front/packages/ui/src/details/collection.tsx
@@ -28,7 +28,7 @@ import {
 import { Container, H2, ImageBackground, Link, P, focusReset, ts } from "@kyoo/primitives";
 import { useTranslation } from "react-i18next";
 import { Theme, useYoshiki } from "yoshiki/native";
-import { ErrorView } from "../fetch";
+import { ErrorView } from "../errors";
 
 export const PartOf = ({
 	name,
