+++ front/translations/en.json
@@ -189,11 +189,12 @@
 	},
 	"errors": {
 		"connection": "Could not connect to the kyoo's server",
-		"connection-tips": "Troublshotting tips:\n - Are you connected to internet?\n - Is your kyoo's server online?\n - Have your account been banned?",
+		"connection-tips": "Troubleshooting tips:\n - Are you connected to internet?\n - Is your kyoo's server online?\n - Have your account been banned?",
 		"unknown": "Unknown error",
 		"try-again": "Try again",
 		"re-login": "Re login",
-		"offline": "You are not connected to internet. Try again later."
+		"offline": "You are not connected to internet. Try again later.",
+		"unauthorized": "You are missing the permissions {{permission}} to access this page."
 	},
 	"mediainfo": {
 		"file": "File",
